requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
celery==5.5.3
redis==6.4.0
librouteros==3.4.1
psycopg2-binary==2.9.9
django-cors-headers==4.6.0

